bpf: use per-cpu scratch space from xdp context to store meta data
Moving to per-CPU scratch space instead avoids cache-misses and allows
to keep the meta buffer always hot in cache. Only really adjust the
pkt meta data space if we push up the stack to signal tc layer that we
handled the service xlation. This improves the performance on my local
testing by roughly +1.2Mpps. We don't perform an expensive clearing of
the scratch space upon entering XDP as usage here is all self-contained
and across tail calls we only read out what we populated beforehand.
