+++ bpf/include/bpf/api.h
@@ -16,23 +16,6 @@
 #include "verifier.h"
 #include "tailcall.h"
 #include "errno.h"
-
-#define PIN_NONE		0
-#define PIN_OBJECT_NS		1
-#define PIN_GLOBAL_NS		2
-
-struct bpf_elf_map {
-	__u32 type;
-	__u32 size_key;
-	__u32 size_value;
-	__u32 max_elem;
-	__u32 flags;
-	__u32 id;
-	__u32 pinning;
-#ifdef SOCKMAP
-	__u32 inner_id;
-	__u32 inner_idx;
-#endif
-};
+#include "loader.h"
 
 #endif /* __BPF_API__ */
